# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  version: 1.0.0
paths:
  /tasks:
    get:
      summary: タスクの一覧を取得
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [NotStarted, InProgress, Completed]
        - name: name
          in: query
          schema:
            type: string
        - name: description
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: sort
          in: query
          schema:
            type: string
            enum: [priority, end_date]
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                  total:
                    type: integer
    post:
      summary: タスクを作成
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: タスク作成成功
...
  /labels:
    get:
      summary: ラベル一覧を取得
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  labels:
                    type: array
                    items:
                      $ref: "#/components/schemas/Label"
    post:
      summary: 新しいラベルを作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelInput"
      responses:
        "201":
          description: 作成成功
...
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        priority:
          type: string
          enum: [High, Middle, Low]
        status:
          type: string
          enum: [NotStarted, InProgress, Completed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
    TaskInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        priority:
          type: string
          enum: [High, Middle, Low]
        status:
          type: string
          enum: [NotStarted, InProgress, Completed]
    Label:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LabelInput:
      type: object
      required:
        - name
        - color
      properties:
        name:
          type: string
        color:
          type: string